got basic font rendering working for the glfw environment.
This also came with the standardization of 0,0 being the top left (see readme)

Moved the shaders to a better spot when operating the software out of project folder (which is the standard)

Note, I don't know if the normal library builds at all right now. All i know is that the test_glfw.cpp is working so I can move on to other things.

Need to formalize a positioning system (floats? ints? doubles?)

Need to formalize vectors too (glm library is helpful but I don't want to export that in the C headers, need something easier, probably just need to make my own)